Move populate_temp_dir test helper to src/test_lib.rs
